backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-dotenv
